+++ lib/util/extensions/player/player_extensions.dart
@@ -19,6 +19,25 @@ extension PlayerExtensions on Player {
     );
   }
 
+  /// This method we notify when detect the enemy when enter in Arc configuration
+  /// Method that bo used in [update] method.
+  void seeEnemyDirectional({
+    required Function(List<Enemy>) observed,
+    VoidCallback? notObserved,
+    double radiusVision = 32,
+    double? angleVision,
+    double angle = 3.14159,
+  }) {
+    if (isDead) return;
+    this.seeComponentsDirectionalByAngleType<Enemy>(
+      observed: observed,
+      notObserved: notObserved,
+      radiusVision: radiusVision,
+      angleVision: angleVision,
+      angle: lastDirection.toRadians(),
+    );
+  }
+
   void simpleAttackMelee({
     Future<SpriteAnimation>? animationRight,
     Future<SpriteAnimation>? animationDown,