+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [next]
+- adds `seeComponentDirectionalByAngle` in `GameComponentExtensions`.
+- adds `seeComponentsDirectionalByAngleType` in `GameComponentExtensions`.
+- adds `seePlayerDirectional` in `NpcExtensions`(Ally and Enemy).
+- adds `seeEnemyDirectional` in `PlayerExtensions`.
+
 # [2.6.4]
 - Update `flame` to `1.2.1`.
 - Update `tiledjsonreader` to `1.1.2`